@@ -7,7 +7,9 @@ import 'package:connection_repository/connection_repository.dart';
 import 'package:equatable/equatable.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:game_domain/game_domain.dart';
-import 'package:match_maker_repository/match_maker_repository.dart' as repo;
+import 'package:match_maker_repository/match_maker_repository.dart';
+import 'package:top_dash/audio/audio_controller.dart';
+import 'package:top_dash/gen/assets.gen.dart';
 import 'package:top_dash_ui/top_dash_ui.dart';
 
 part 'game_event.dart';
@@ -16,7 +18,8 @@ part 'game_state.dart';
 class GameBloc extends Bloc<GameEvent, GameState> {
   GameBloc({
     required GameResource gameResource,
-    required repo.MatchMakerRepository matchMakerRepository,
+    required MatchMakerRepository matchMakerRepository,
+    required AudioController audioController,
     required MatchSolver matchSolver,
     required User user,
     required this.isHost,
@@ -24,6 +27,7 @@ class GameBloc extends Bloc<GameEvent, GameState> {
   })  : _gameResource = gameResource,
         _matchMakerRepository = matchMakerRepository,
         _connectionRepository = connectionRepository,
+        _audioController = audioController,
         _matchSolver = matchSolver,
         _user = user,
         super(const MatchLoadingState()) {
@@ -40,11 +44,13 @@ class GameBloc extends Bloc<GameEvent, GameState> {
   }
 
   final GameResource _gameResource;
-  final repo.MatchMakerRepository _matchMakerRepository;
+  final MatchMakerRepository _matchMakerRepository;
   final MatchSolver _matchSolver;
   final User _user;
   final bool isHost;
+  final AudioController _audioController;
   final ConnectionRepository _connectionRepository;
+  final List<String> playedCardsInOrder = [];
   Timer? _turnTimer;
 
   // Added to easily toggle timer functionality for testing purposes.
@@ -52,7 +58,7 @@ class GameBloc extends Bloc<GameEvent, GameState> {
   static const _turnMaxTime = 10;
 
   StreamSubscription<MatchState>? _stateSubscription;
-  StreamSubscription<repo.Match>? _opponentDisconnectSubscription;
+  StreamSubscription<DraftMatch>? _opponentDisconnectSubscription;
   StreamSubscription<ScoreCard>? _scoreSubscription;
 
   Future<void> _onMatchRequested(
@@ -75,11 +81,12 @@ class GameBloc extends Bloc<GameEvent, GameState> {
       if (match == null || matchState == null || scoreCard == null) {
         emit(const MatchLoadFailedState());
       } else {
+        _audioController.playSfx(Assets.sfx.startGame);
         emit(
           MatchLoadedState(
             match: match,
             matchState: matchState,
-            turns: const [],
+            rounds: const [],
             playerScoreCard: scoreCard,
             turnTimeRemaining: _turnMaxTime,
             turnAnimationsFinished: true,
@@ -120,35 +127,44 @@ class GameBloc extends Bloc<GameEvent, GameState> {
           ? event.updatedState.guestPlayedCards
           : event.updatedState.hostPlayedCards;
 
-      final isPlayerMove = matchLoadedState.turns
-              .where((turn) => turn.playerCardId != null)
-              .length !=
-          matchStatePlayerMoves.length;
+      String? lastPlayedCard;
+      for (final cardId in [
+        ...matchStatePlayerMoves,
+        ...matchStateOpponentMoves
+      ]) {
+        if (!playedCardsInOrder.contains(cardId)) {
+          playedCardsInOrder.add(cardId);
+          lastPlayedCard = cardId;
+        }
+      }
 
       final moveLength = math.max(
         matchStatePlayerMoves.length,
         matchStateOpponentMoves.length,
       );
 
-      final turns = [
+      final rounds = [
         for (var i = 0; i < moveLength; i++)
-          MatchTurn(
+          MatchRound(
             playerCardId: i < matchStatePlayerMoves.length
                 ? matchStatePlayerMoves[i]
                 : null,
             opponentCardId: i < matchStateOpponentMoves.length
                 ? matchStateOpponentMoves[i]
                 : null,
-            showCardsOverlay: i < matchLoadedState.turns.length &&
-                matchLoadedState.turns[i].showCardsOverlay,
+            showCardsOverlay: i < matchLoadedState.rounds.length &&
+                matchLoadedState.rounds[i].showCardsOverlay,
           ),
       ];
 
+      if (lastPlayedCard != null) {
+        _audioController.playSfx(Assets.sfx.playCard);
+      }
       emit(
         matchLoadedState.copyWith(
           matchState: event.updatedState,
-          turns: turns,
-          playerPlayed: isPlayerMove ? false : null,
+          rounds: rounds,
+          lastPlayedCardId: lastPlayedCard,
         ),
       );
 
@@ -179,7 +195,6 @@ class GameBloc extends Bloc<GameEvent, GameState> {
       final matchState = state as MatchLoadedState;
       emit(
         matchState.copyWith(
-          playerPlayed: true,
           turnAnimationsFinished: false,
         ),
       );
@@ -243,7 +258,10 @@ class GameBloc extends Bloc<GameEvent, GameState> {
   ) {
     if (state is MatchLoadedState) {
       final matchLoadedState = state as MatchLoadedState;
-      if (isPlayerTurn && !matchLoadedState.matchState.isOver()) {
+      if (isPlayerAllowedToPlay &&
+          !matchLoadedState.matchState.isOver() &&
+          matchLoadedState.matchState.hostPlayedCards.length ==
+              matchLoadedState.matchState.guestPlayedCards.length) {
         emit(matchLoadedState.copyWith(turnTimeRemaining: _turnMaxTime));
 
         _turnTimer = Timer.periodic(
@@ -315,12 +333,12 @@ class GameBloc extends Bloc<GameEvent, GameState> {
   ) {
     if (state is MatchLoadedState) {
       final matchLoadedState = state as MatchLoadedState;
-      if (matchLoadedState.turns.isNotEmpty) {
-        final lastTurn = matchLoadedState.turns.last;
-        final turns = [...matchLoadedState.turns]
+      if (matchLoadedState.rounds.isNotEmpty) {
+        final lastRound = matchLoadedState.rounds.last;
+        final rounds = [...matchLoadedState.rounds]
           ..removeLast()
-          ..add(lastTurn.copyWith(showCardsOverlay: true));
-        emit(matchLoadedState.copyWith(turns: turns));
+          ..add(lastRound.copyWith(showCardsOverlay: true));
+        emit(matchLoadedState.copyWith(rounds: rounds));
       }
     }
   }
@@ -339,7 +357,7 @@ class GameBloc extends Bloc<GameEvent, GameState> {
           .indexWhere((id) => id == card.id);
 
       if (round >= 0) {
-        final turn = matchLoadedState.turns[round];
+        final turn = matchLoadedState.rounds[round];
         if (turn.isComplete() && turn.showCardsOverlay) {
           final result = _matchSolver.calculateRoundResult(
             matchLoadedState.match,
@@ -365,10 +383,10 @@ class GameBloc extends Bloc<GameEvent, GameState> {
     return null;
   }
 
-  bool get isPlayerTurn {
+  bool get isPlayerAllowedToPlay {
     if (state is MatchLoadedState) {
       final matchLoadedState = state as MatchLoadedState;
-      return _matchSolver.isPlayerTurn(
+      return _matchSolver.isPlayerAllowedToPlay(
         matchLoadedState.matchState,
         isHost: isHost,
       );
@@ -420,23 +438,6 @@ class GameBloc extends Bloc<GameEvent, GameState> {
     return [];
   }
 
-  String? get lastPlayedCardId {
-    if (state is MatchLoadedState) {
-      final matchLoadedState = state as MatchLoadedState;
-      if (matchLoadedState.turns.isNotEmpty) {
-        final lastTurn = matchLoadedState.turns.last;
-        if (lastTurn.isComplete()) {
-          if (isPlayerTurn) {
-            return lastTurn.playerCardId;
-          }
-          return lastTurn.opponentCardId;
-        }
-        return lastTurn.playerCardId ?? lastTurn.opponentCardId;
-      }
-    }
-    return null;
-  }
-
   @override
   Future<void> close() {
     _stateSubscription?.cancel();
@@ -448,7 +449,7 @@ class GameBloc extends Bloc<GameEvent, GameState> {
   }
 }
 
-extension MatchTurnX on MatchTurn {
+extension MatchTurnX on MatchRound {
   bool isComplete() {
     return playerCardId != null && opponentCardId != null;
   }
@@ -456,7 +457,7 @@ extension MatchTurnX on MatchTurn {
 
 extension MatchLoadedStateX on MatchLoadedState {
   bool isCardTurnComplete(Card card) {
-    for (final turn in turns) {
+    for (final turn in rounds) {
       if (card.id == turn.playerCardId || card.id == turn.opponentCardId) {
         return turn.isComplete();
       }
