@@ -335,7 +335,6 @@ void main() {
           matchId: 'matchId',
           guestPlayedCards: ['a'],
           hostPlayedCards: ['b'],
-          hostStartsMatch: true,
         );
 
         when(() => response.statusCode).thenReturn(HttpStatus.ok);
@@ -486,6 +485,32 @@ void main() {
             ),
           ),
         );
+      });
+
+      group('shareHandUrl', () {
+        test('returns the correct url', () {
+          when(
+            () => apiClient.shareHandUrl(any()),
+          ).thenReturn('http://baseurl.com/public/share?deckId=id');
+
+          expect(
+            resource.shareHandUrl('id'),
+            equals('http://baseurl.com/public/share?deckId=id'),
+          );
+        });
+      });
+
+      group('shareCardUrl', () {
+        test('returns the correct url', () {
+          when(
+            () => apiClient.shareCardUrl(any()),
+          ).thenReturn('http://baseurl.com/public/share?cardId=id');
+
+          expect(
+            resource.shareCardUrl('id'),
+            equals('http://baseurl.com/public/share?cardId=id'),
+          );
+        });
       });
     });
   });
